test(vgg_architecture): test all the expected layers in the production vgg pretrained model
